fix: remove alunos files and adding routes to app file
