--- database/alunos.py
@@ -1,7 +1,0 @@
-alunos =  [
-        {"id": 1, "nome": "João Pereira", "idade": 15},
-        {"id": 2, "nome": "Mariana Lima", "idade": 14},
-        {"id": 3, "nome": "Lucas Oliveira", "idade": 16},
-        {"id": 4, "nome": "Beatriz Santos", "idade": 15},
-        {"id": 5, "nome": "Gabriel Martins", "idade": 14}
-]
